feat: pipeline para ingestão e atualização

+++ src/analytics/exec_query.py
@@ -25,7 +25,7 @@ def date_range(start, stop, monthly=False):
     
     return dates
 
-def exec_query(table, db_origin, db_target, dt_start, dt_stop, monthly=False):
+def exec_query(table, db_origin, db_target, dt_start, dt_stop, monthly=False, mode='append'):
     engine_app = sqlalchemy.create_engine(f'sqlite:///../../data/{db_origin}/database.db')
     engine_analytical = sqlalchemy.create_engine(f'sqlite:///../../data/{db_target}/database.db')
 
@@ -33,17 +33,18 @@ def exec_query(table, db_origin, db_target, dt_start, dt_stop, monthly=False):
     dates = date_range(dt_start, dt_stop, monthly)
 
     for i in tqdm(dates):
-        with engine_analytical.connect() as con:
-            try:
-                query_delete = f"DELETE FROM {table} WHERE dtRef = date('{i}', '-1 day')"
-                con.execute(sqlalchemy.text(query_delete))
-                con.commit()
-            except Exception as err:
-                print(err)
+        if mode == 'append':
+            with engine_analytical.connect() as con:
+                try:
+                    query_delete = f"DELETE FROM {table} WHERE dtRef = date('{i}', '-1 day')"
+                    con.execute(sqlalchemy.text(query_delete))
+                    con.commit()
+                except Exception as err:
+                    print(err)
 
         query_format = query.format(date=i)
         df = pd.read_sql(query_format, engine_app)
-        df.to_sql(f'{table}', engine_analytical, index=False, if_exists='append') 
+        df.to_sql(f'{table}', engine_analytical, index=False, if_exists=mode) 
 
 def main():
     
@@ -56,10 +57,11 @@ def main():
     parser.add_argument('--start', type=str, default='2024-03-01')   # Data inicial do processamento
     parser.add_argument('--stop', type=str, default=now)    # Data final do processamento
     parser.add_argument('--monthly', action='store_true')   # Processar apenas o 1o dia do mês
+    parser.add_argument('--mode', type=str, choices=['append', 'replace'], default='append')   # Modo de inserção de dados
 
     args = parser.parse_args()
 
-    exec_query(args.table, args.db_origin, args.db_target, args.start, args.stop, args.monthly)
+    exec_query(args.table, args.db_origin, args.db_target, args.start, args.stop, args.monthly, args.mode)
 
 if __name__ == '__main__':
     main()
